+++ backup/init/2-table.sql
@@ -1,0 +1,3 @@
+/* CREATE TABLE tb1(
+  name VARCHAR(20)
+); */