+++ cli/main/shell/lnmp-docker.sh
@@ -1,0 +1,1 @@
+../../../lnmp-docker.sh