+++ config/update.sh
@@ -6,6 +6,8 @@ sed -i "" "s#^logfile.*#logfile /var/log/redis/redis.log#g" redis/redis.conf
 
 sed -i "" "s/^dir.*/dir \/data/g" redis/redis.conf
 
+sed -i '' 's#^bind.*#bind 0.0.0.0#g' redis/redis.conf
+
 cd php
 
 rm -rf *.ini