+++ dockerfile/update.sh
@@ -1,5 +1,3 @@
 #!/usr/bin/env bash
 
 cp ../config/redis/redis.conf redis/
-
-cp ../config/etc/nginx/nginx.conf nginx/nginx.conf